fix: check on tAGENCY_CLASSIFICATION for filtering, test on right file
